Update Spinners and CheckBoxes to clean up automatically
setChildrenEnabled is called whenever the state changes between open, started, paused, stopped, and closed. When the stream is closed, all the spinners and checkboxes should work as expected. Otherwise, disable all of them.

This change fixes the user complaints about MMAP not working as expected when the checkbox is toggled by removing the ability to toggle the checkbox