Added mDeviceId to FilterAudioStream (#1310)
Fixed a missing parameter in FilterAudioStream. Without this change, when getDeviceId() is called, the device id is set to 0, as it uses the default value. With this change, mDeviceId correctly uses the value from the child stream.

I verified that the issue repros before the change and the fix does indeed fix the problem